+++ armory/eval_scripts/carlini_wagner_attack_targeted.py
@@ -11,7 +11,7 @@ import sys
 
 import numpy as np
 
-from armory.data.utils import load_dataset
+from armory.utils.config_loading import load_dataset
 from armory.eval import plot
 from armory.paths import DockerPaths
 
@@ -66,8 +66,8 @@ def evaluate_classifier(config_path: str) -> None:
     preprocessing_fn = getattr(classifier_module, "preprocessing_fn")
 
     logger.info(f"Loading dataset {config['dataset']['name']}...")
-    x_train, y_train, x_test, y_test = datasets.load(
-        config["dataset"]["name"], preprocessing_fn=preprocessing_fn
+    x_train, y_train, x_test, y_test = load_dataset(
+        config["dataset"], preprocessing_fn=preprocessing_fn
     )
 
     logger.info(
