+++ examples/fgm_attack_binary_search.json
@@ -20,7 +20,8 @@
         "name": "FGMBinarySearch"
     },
     "dataset": {
-        "name": "mnist"
+        "module": "armory.data.datasets",
+        "name": "mnist_data"
     },
     "defense": null,
     "evaluation": {
