+++ examples/transform_defense.json
@@ -4,6 +4,7 @@
     },
     "attack": null,
     "dataset": {
+        "module": "armory.data.datasets",
         "name": "imagenet_adversarial"
     },
     "defense": {
